pyproject.toml
[build-system]
requires = ["setuptools>=65"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["axswarm"]

[project]
name = "axswarm"
version = "0.1.0"
license = "MIT"
readme = "README.md"
authors = [{ name = "Martin Schuck" }]
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
]
keywords = ["drones", "control", "swarm", "path planning"]
requires-python = ">=3.8"
dependencies = [
    "jax>=0.6.0",
    "flax>=0.10.6",
    "numpy>=2.2.0",
    "einops>=0.8.0",
]

[project.optional-dependencies]
examples = [
    "fire>=0.7.0",
    "matplotlib>=3.10.0",
    "pyyaml>=6.0.2",
    "crazyflow[gpu]>=0.0.1",
]


[tool.ruff]
line-length = 100
target-version = "py313"
select = ["E", "F", "B", "I", "N", "UP", "RUF"]
ignore = ["N803", "N806"]


[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
axswarm = { path = ".", editable = true}
crazyflow = { version = ">=0.0.1", extras = ["gpu"]}

[tool.pixi.tasks]

[tool.pixi.dependencies]
numpy = ">=2.3.0,<3"
matplotlib = ">=3.10.3,<4"
pyyaml = ">=6.0.2,<7"
scipy = ">=1.15.2,<2"
fire = ">=0.7.0,<0.8"
pulp = ">=2.8.0,<3"
uv = ">=0.7.16,<0.8"
pyinstrument = ">=5.0.2,<6"

